Move include to include/concurrentqueue and adapt makefiles
Adapt makefiles, cpp to match new include/concurrentqueue include path
