+++ tests/fuzztests/fuzztests.cpp
@@ -29,7 +29,7 @@
 #include <windows.h>
 #endif
 
-#include "../../concurrentqueue.h"
+#include "concurrentqueue/concurrentqueue.h"
 #include "../common/simplethread.h"
 #include "../common/systemtime.h"
 #include "../corealgos.h"
