+++ tests/unittests/mallocmacro.cpp
@@ -1,4 +1,4 @@
 #define malloc(x) malloc(x)
 #define free(x) free(x)
 
-#include "../../blockingconcurrentqueue.h"
+#include "concurrentqueue//blockingconcurrentqueue.h"
