+++ tests/unittests/unittests.cpp
@@ -24,8 +24,8 @@
 #include "minitest.h"
 #include "../common/simplethread.h"
 #include "../common/systemtime.h"
-#include "../../concurrentqueue.h"
-#include "../../blockingconcurrentqueue.h"
+#include "concurrentqueue/concurrentqueue.h"
+#include "concurrentqueue/blockingconcurrentqueue.h"
 
 namespace {
 	struct tracking_allocator
